Added an example of bootstrapping custom build system based on Blueprint.
The example was originally written by Tim Kilbourn.

It was necessary to modify it for the latest Blueprint; in particular,
for the minibp-based bootstrapping process (as opposed to the former
process based on prebuilt build.ninja.in file). Dependency handling
has also changed. Currently the example is not "idiomatic Blueprint"
and should be modernized further in future.