Allow cloned modules to propagate arbitrary data through immutable pointers
Support propagating arbitrary data between cloned modules by blindly
copying pointers tagged with `blueprint:"immutable_ptr"`.  The type
of the property must be a pointer, and the value referenced to must be
nil or never changed, as the value will be shared by multiple modules.
Unfortunately Go has no way to enforce immutability on a type, so this
has to be on the honor system.

Test: TestCloneProperties, TestCloneEmptyProperties, TestZeroProperties, TestExtendProperties
Change-Id: I488de0005c7883e5180b7ffeb2d69e4d1f9aea44